# Batch Insert and Batch Delete now do ten records per query