--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,4 @@ docs/build
 coverage.xml
 .coverage
 test.py
+experiment.py