--- a/tests/test_airtable_class.py
+++ b/tests/test_airtable_class.py
@@ -23,6 +23,14 @@ def test_url(base_key, table_name, table_url_suffix):
     assert table.url_table == "{0}/{1}".format(table.API_URL, table_url_suffix)
 
 
+def test_chunk(table):
+    chunks = [chunk for chunk in table._chunk([0, 0, 1, 1, 2, 2, 3], 2)]
+    assert chunks[0] == [0, 0]
+    assert chunks[1] == [1, 1]
+    assert chunks[2] == [2, 2]
+    assert chunks[3] == [3]
+
+
 def test_record_url(table):
     rv = table.record_url("xxx")
     assert rv == urljoin(table.url_table, "xxx")
@@ -117,12 +125,11 @@ def test_search_not_found(table, mock_response_single):
 
 def test_batch_insert(table, mock_records):
     with Mocker() as mock:
-        for record in mock_records:
+        for chunk in _chunk(mock_records, 10):
             mock.post(
                 table.url_table,
                 status_code=201,
-                json=record,
-                additional_matcher=match_request_data(record["fields"]),
+                json={'records': chunk},
             )
         records = [i["fields"] for i in mock_records]
         resp = table.batch_insert(records)
@@ -227,19 +234,25 @@ def test_delete_by_field(table, mock_response_single):
 def test_batch_delete(table, mock_records):
     ids = [i["id"] for i in mock_records]
     with Mocker() as mock:
-        for id_ in ids:
+        for chunk in _chunk(ids, 10):
+            params = [('records', id_) for id_ in chunk]
+            params_encode = urlencode(params)
             mock.delete(
-                urljoin(table.url_table, id_),
+                table.url_table + '?' + params_encode,
                 status_code=201,
-                json={"delete": True, "id": id_},
+                json={'records': [{"delete": True, "id": id_} for id_ in chunk]},
             )
-        records = [i["id"] for i in mock_records]
-        resp = table.batch_delete(records)
+
+        resp = table.batch_delete(ids)
     expected = [{"delete": True, "id": i} for i in ids]
     assert resp == expected
 
 
 # Helpers
+
+def _chunk(iterable, chunk_size):
+    for i in range(0, len(iterable), chunk_size):
+        yield iterable[i:i + chunk_size]
 
 
 def match_request_data(post_data):