Hi! 

I've updated the airtable class to perform batch updates for insert and delete. For large sets of records this approach is _much_ faster (~factor of 10, see below) than one-request-per-item due to the Airtable API request limit. 

While the previous implementation used the `_batch_request` helper function, I've inlined the logic for `batch_insert` and `batch_delete` as they utilise POST and query strings respectively so I thought best to keep them separate. 

I've updated the tests to reflect the changes to these two methods and they are passing for all Python versions (2.7, 3.5, 3.6, 3.7, 3.8)

**Benchmarking:**
New implementation
`Inserted 100 records in 4.0554 seconds
Deleted 100 records in 3.6701 seconds`

Old implmentation
`Inserted 100 records in 38.4023 seconds
Deleted 100 records in 36.7222 seconds`

Average improvement factor: ~9.7x!